Add callback to expandBlockList()
Since the entire point of runnning expandBlockList() is to allow the usernames to be scraped from the page, and the function is timer based, I added a callback to make it easier for functions to run immediately after expandBlockList().

This does make returning values from functions used as a callback more difficult.