Add popup with copyable blacklist
When using the Export Blacklist feature, a popup will appear that allows you to copy a comma separated string of all of the blogs you are blocking.

This is a bit hacky, but I wanted to roll out this feature quickly, and I don't know how long it will take me to learn how to do it neatly. As it is: it works, and fairly quickly.