Clean list of users getting blocked
Prevents duplicates both between the list to be blocked and those already blocked and within the list itself.
This is necessary for a planned fix. That fix will check the last blocked user to verify that it was successful. Removing duplicates will add some up-front time (more for users who have many blocks), but will greatly reduce the time taken to ensure proper blocks (planning a timeout feature).