Convert to looping with MutationObserver
Reusing code from export.js to extract an array of currently blocked users.
Filtering the user input so that only users that aren't already blocked will be handled.
Remove the setTimeouts and setIntervals (no longer needed, because next block is triggered by the completion of the previous block).

No longer using openBlockList, because it is already handled by expandBlockList.

This is MUCH FASTER, and I tested it against more than 500 users. I probably need to test it on a very slow computer, but this should make computer speed irrelevant for accuracy.